--- packages/hash/api/src/graphql/typeDefs/entityType.typedef.ts
@@ -1,181 +1,0 @@
-import { gql } from "apollo-server-express";
-
-export const deprecatedEntityTypeTypedef = gql`
-  extend type Query {
-    """
-    Get an entity type by its fixed id, which is:
-    - entityId on the EntityType itself, or
-    - entityTypeId on an Entity of its type)
-    Or, by EntityTypeChoice
-    """
-    deprecatedGetEntityType(
-      entityTypeId: ID
-        @deprecated(
-          reason: """
-          Use "choice.entityTypeId" instead
-          """
-        )
-      choice: EntityTypeChoice
-    ): DeprecatedEntityType!
-    """
-    Get all EntityTypes belonging to an account.
-    Optionally include:
-    - types in use by the account, but belonging to other accounts.
-    - ALL types.
-    TODO: replace this with a new aggregateEntityTypes query instead
-    """
-    deprecatedGetAccountEntityTypes(
-      accountId: ID!
-      includeAllTypes: Boolean = false
-      includeOtherTypesInUse: Boolean = false
-    ): [DeprecatedEntityType!]!
-  }
-
-  extend type Mutation {
-    """
-    Create an entity type
-    """
-    deprecatedCreateEntityType(
-      accountId: ID!
-      """
-      The name for the type. Must be unique in the given account.
-      """
-      name: String!
-      """
-      A description for the type.
-      """
-      description: String
-      """
-      The schema definition for the entity type, in JSON Schema.
-      """
-      schema: JSONObject
-    ): DeprecatedEntityType!
-
-    """
-    Update an entity type
-    """
-    deprecatedUpdateEntityType(
-      """
-      The fixed id of the entityType to update, i.e. its entityId
-      """
-      entityId: ID!
-      """
-      The schema definition for the entity type, in JSON Schema format.
-      """
-      schema: JSONObject!
-    ): DeprecatedEntityType!
-  }
-
-  enum SystemTypeName {
-    Block
-    Comment
-    EntityType
-    Org
-    Page
-    Text
-    User
-    OrgMembership
-    File
-    OrgInvitationLink
-    OrgEmailInvitation
-  }
-
-  """
-  A schema describing and validating a specific type of entity
-  """
-  type DeprecatedEntityType implements Entity {
-    """
-    The shape of the entity, expressed as a JSON Schema
-    https://json-schema.org/
-    """
-    properties: JSONObject!
-
-    # ENTITY INTERFACE FIELDS BEGIN #
-    """
-    The id of the entity - alias of 'entityId'
-    """
-    id: ID!
-    """
-    The id of the entity - alias of 'id'
-    """
-    entityId: ID!
-    """
-    The specific version if of the entity
-    """
-    entityVersionId: ID!
-    """
-    The id of the account this entity belongs to
-    """
-    accountId: ID!
-    """
-    The date the entity was created
-    """
-    createdAt: Date!
-    """
-    The date this entity version was created.
-    """
-    entityVersionCreatedAt: Date!
-    """
-    The user who created the entity
-    """
-    createdByAccountId: ID!
-    """
-    The date the entity was last updated
-    """
-    updatedAt: Date!
-    """
-    The visibility level of the entity
-    """
-    visibility: Visibility!
-    """
-    The fixed id of the type this entity is of
-    """
-    entityTypeId: ID!
-    """
-    The id of the specific version of the type this entity is of
-    """
-    entityTypeVersionId: ID!
-    """
-    The name of the entity type this belongs to.
-    N.B. Type names are unique by account - not globally.
-    """
-    entityTypeName: String!
-    """
-    The full entityType definition
-    """
-    entityType: DeprecatedEntityType!
-    """
-    The version timeline of the entity.
-    """
-    history: [EntityVersion!]
-    """
-    The outgoing links of the entity.
-    """
-    linkGroups: [LinkGroup!]!
-    """
-    The linked entities of the entity.
-    """
-    linkedEntities: [UnknownEntity!]!
-    """
-    The linked aggregations of the entity.
-    """
-    linkedAggregations: [LinkedAggregation!]!
-    # ENTITY INTERFACE FIELDS END #
-
-    """
-    Retrieve all EntityTypes that are children of
-    the current EntityType.
-    """
-    children: [DeprecatedEntityType!]
-    """
-    Retrieve all EntityTypes that are parents of
-    the current EntityType.
-    """
-    parents: [DeprecatedEntityType!]
-
-    """
-    Retrieve all ancestors of the current Entity Type (resolving the parents, parents' parents and so forth)
-    """
-    ancestors: [DeprecatedEntityType!]
-  }
-`;
