+++ packages/hash/frontend/src/lib/entities.ts
@@ -1,11 +1,4 @@
-import {
-  Entity as BpEntity,
-  EntityType as BpEntityType,
-  Link as BpLink,
-  LinkGroup as BpLinkGroup,
-  LinkedAggregation as BpLinkedAggregation,
-  LinkedAggregationDefinition as BpLinkedAggregationDefinition,
-} from "@blockprotocol/graph";
+import { Entity as BpEntity, Link as BpLink } from "@blockprotocol/graph";
 
 import {
   extractEntityUuidFromEntityId,
@@ -16,10 +9,7 @@ import { getRoots } from "@hashintel/hash-subgraph/src/stdlib/roots";
 import { getPropertyTypesByBaseUri } from "@hashintel/hash-subgraph/src/stdlib/element/property-type";
 import {
   UnknownEntity as ApiEntity,
-  DeprecatedEntityType as ApiEntityType,
   Link as ApiLink,
-  LinkGroup as ApiLinkGroup,
-  LinkedAggregation as ApiLinkedAggregation,
 } from "../graphql/apiTypes.gen";
 
 const isObject = (thing: unknown): thing is {} =>
@@ -264,16 +254,6 @@ type ApiLinkedAggregationIdentifier = {
 };
 
 /**
- * Create a aggregationId that's a stringified object containing the fields we will need later.
- * The re-written aggregationId is what should be sent to the block with any Link
- */
-const rewriteLinkedAggregationIdentifier = ({
-  accountId,
-  aggregationId,
-}: ApiLinkedAggregationIdentifier) =>
-  JSON.stringify({ accountId, aggregationId });
-
-/**
  * We send blocks an 'aggregationId' that is a stringified object in {@link convertApiLinkedAggregationToBpLinkedAggregation}
  * – this reverses the process so we have the individual fields to use in calling the HASH API.
  *
@@ -321,124 +301,6 @@ export const parseLinkedAggregationIdentifier = (
 
   return identifierObject;
 };
-
-export const convertApiLinkGroupToBpLinkGroup = (
-  linkGroup: ApiLinkGroup,
-): BpLinkGroup => {
-  const { sourceAccountId, sourceEntityId, links, ...rest } = linkGroup;
-
-  if (sourceEntityId.includes("{")) {
-    throw new Error(
-      `sourceEntityId has already been re-written as a stringified object: ${sourceEntityId}`,
-    );
-  }
-
-  return {
-    sourceEntityId: rewriteEntityIdentifier({
-      accountId: sourceAccountId,
-      entityId: sourceEntityId,
-    }),
-    links: convertApiLinksToBpLinks(links),
-    ...rest,
-  };
-};
-
-export const convertApiLinkGroupsToBpLinkGroups = (
-  linkGroups: ApiLinkGroup[],
-): BpLinkGroup[] =>
-  linkGroups.map((record) => convertApiLinkGroupToBpLinkGroup(record));
-
-type MinimalApiLinkedAggregation = Pick<
-  ApiLinkedAggregation,
-  "aggregationId" | "sourceAccountId" | "sourceEntityId" | "operation" | "path"
->;
-
-export function convertApiLinkedAggregationToBpLinkedAggregation(
-  linkedAggregation: MinimalApiLinkedAggregation & {
-    results: Pick<
-      ApiEntity,
-      "accountId" | "entityId" | "entityTypeId" | "properties"
-    >[];
-  },
-): BpLinkedAggregation;
-
-export function convertApiLinkedAggregationToBpLinkedAggregation(
-  linkedAggregation: MinimalApiLinkedAggregation,
-): BpLinkedAggregationDefinition;
-
-/**
- * Converts a LinkedAggregation from its GraphQL API representation to its Block Protocol representation,
- * by re-writing all of aggregationId, sourceEntityId and sourceDestinationId so that they are
- * stringified objects containing the identifiers we need (i.e. to include accountId)
- */
-export function convertApiLinkedAggregationToBpLinkedAggregation(
-  linkedAggregation:
-    | (MinimalApiLinkedAggregation & {
-        results: Pick<
-          ApiEntity,
-          "accountId" | "entityId" | "entityTypeId" | "properties"
-        >[];
-      })
-    | MinimalApiLinkedAggregation,
-): BpLinkedAggregation | BpLinkedAggregationDefinition {
-  const { aggregationId, sourceAccountId, sourceEntityId, operation, path } =
-    linkedAggregation;
-  if (aggregationId.includes("{")) {
-    throw new Error(
-      `aggregationId has already been re-written as a stringified object: ${aggregationId}`,
-    );
-  }
-
-  if (sourceEntityId.includes("{")) {
-    throw new Error(
-      `sourceEntityId has already been re-written as a stringified object: ${sourceEntityId}`,
-    );
-  }
-
-  return {
-    aggregationId: rewriteLinkedAggregationIdentifier({
-      accountId: sourceAccountId,
-      aggregationId,
-    }),
-    sourceEntityId: rewriteEntityIdentifier({
-      accountId: sourceAccountId,
-      entityId: sourceEntityId,
-    }),
-    path,
-    operation: operation as BpLinkedAggregation["operation"],
-    results:
-      "results" in linkedAggregation && linkedAggregation.results
-        ? convertApiEntitiesToBpEntities(linkedAggregation.results)
-        : undefined,
-  };
-}
-
-/**
- * Converts an entity type from its GraphQL API representation to its Block Protocol representation:
- * 1. Only provide 'entityTypeId' at the top level
- * 2. Provide the schema under 'schema', not 'properties'
- *
- * N.B. this intentionally does not re-write 'entityTypeId' to include accountId, since types are not sharded,
- * and the 'entityTypeId' is sufficient to identify entity types when calling the HASH API.
- */
-export const convertApiEntityTypeToBpEntityType = ({
-  entityId,
-  properties,
-}: Pick<ApiEntityType, "entityId" | "properties">): BpEntityType => {
-  return {
-    entityTypeId: entityId,
-    schema: properties,
-  };
-};
-
-/**
- * Converts entity types from their GraphQL API representation to their Block Protocol representation
- * @see convertApiEntityTypeToBpEntityType
- */
-export const convertApiEntityTypesToBpEntityTypes = (
-  records: Pick<ApiEntityType, "entityId" | "properties">[],
-): BpEntityType[] =>
-  records.map((record) => convertApiEntityTypeToBpEntityType(record));
 
 /**
  * Generate a display label for an entity
