+++ packages/hash/shared/src/graphql/createApolloClient.ts
@@ -13,7 +13,6 @@ import {
 } from "node-fetch";
 
 import { apiGraphQLEndpoint } from "../environment";
-import { SystemTypeName } from "./apiTypes.gen";
 
 import possibleTypes from "./fragmentTypes.gen.json";
 
@@ -82,25 +81,10 @@ export const createApolloClient = (params?: {
 
   const entityKeyFields = { keyFields: ["entityId"] };
 
-  const typePolicies: Record<SystemTypeName, typeof entityKeyFields> = {
-    Block: entityKeyFields,
-    Comment: entityKeyFields,
-    EntityType: entityKeyFields,
-    OrgEmailInvitation: entityKeyFields,
-    OrgInvitationLink: entityKeyFields,
-    Org: entityKeyFields,
-    OrgMembership: entityKeyFields,
-    Page: entityKeyFields,
-    Text: entityKeyFields,
-    User: entityKeyFields,
-    File: entityKeyFields,
-  };
-
   return new ApolloClient({
     cache: new InMemoryCache({
       possibleTypes: possibleTypes.possibleTypes,
       typePolicies: {
-        ...typePolicies,
         UnknownEntity: entityKeyFields,
         Link: { keyFields: ["linkId"] },
         PageProperties: { keyFields: ["pageEntityId"] },
