+++ packages/hash/shared/src/util.ts
@@ -1,7 +1,6 @@
 // @todo this should be defined elsewhere
 
 import { uniq } from "lodash";
-import { FileProperties } from "./graphql/apiTypes.gen";
 
 export type DistributiveOmit<T, K extends keyof any> = T extends any
   ? Omit<T, K>
@@ -42,15 +41,6 @@ export const collect = <P extends Array<any>>(
       handler(thisCalls);
     });
   };
-};
-
-export const isFileProperties = (props: {}): props is FileProperties => {
-  return (
-    "key" in props &&
-    "size" in props &&
-    "url" in props &&
-    "storageType" in props
-  );
 };
 
 /** A `Map` which creates a default value if the value for a key is not set. */
